refactor(architecture-diagram): make the engine extraction-ready
Stack-Id: architecture-diagram-m4-20260904
Stack-Position: 3/3
