+++ .gitignore
@@ -44,11 +44,8 @@ site/dist/
 *.svg
 
 # Package fixtures and committed reference assets are test inputs, not
-# generated output. The blanket rules above hid a test fixture from every
-# clean checkout, so the suite that reads it passed only on machines where
-# an untracked local copy happened to exist. Scoped to <type>/<package>/ so
-# it covers every package type the test gate iterates, not just skills.
-!*/*/tests/fixtures/*.svg
+# generated output. This supports package-local tests at any nesting depth.
+!**/tests/fixtures/*.svg
 !*/*/assets/*.svg
 !*/*/references/**/*.svg
