+++ skills/architecture-diagram/engine/__init__.py
@@ -1,0 +1,1 @@
+"""Architecture diagram rendering engine."""
