+++ skills/architecture-diagram/engine/__main__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .render import main
+
+raise SystemExit(main())
