+++ skills/architecture-diagram/engine/stencil2svg.py
