+++ skills/architecture-diagram/engine/svg_inline.py
