+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-ambiguous-region.yaml
