+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-inconsistent-security-region.yaml
