+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-missing-boundary-mechanism.yaml
