+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-missing-owner.yaml
