+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-missing-region.yaml
