+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-missing-required-zone-kinds.yaml
