+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-storage-outside-security.yaml
