+++ skills/architecture-diagram/engine/tests/fixtures/deployment-profile-valid.yaml
