+++ skills/architecture-diagram/engine/tests/fixtures/example-serverless.layout.json
