+++ skills/architecture-diagram/engine/tests/fixtures/real_shapes.xml
