+++ skills/architecture-diagram/engine/tests/fixtures/spec-compare-base.yaml
