+++ skills/architecture-diagram/engine/tests/fixtures/spec-compare-head.yaml
