+++ skills/architecture-diagram/engine/tests/fixtures/spec-compare-reordered.yaml
