+++ skills/architecture-diagram/references/editability.md
@@ -1,12 +1,12 @@
 # Why the output is "completely editable" — and what that actually means
 
-"Editable SVG" is easy to claim and easy to get wrong in ways that only surface once someone tries to actually edit the file. This document is the design rationale behind `scripts/render.py`'s output contract, and the concrete, checkable definition `check_editability()` enforces on every render.
+"Editable SVG" is easy to claim and easy to get wrong in ways that only surface once someone tries to actually edit the file. This document is the design rationale behind `engine/render.py`'s output contract, and the concrete, checkable definition `check_editability()` enforces on every render.
 
 ## The three ways an SVG can quietly fail to be editable
 
-**1. Raster icons wrapped in an SVG shell.** The most common failure mode among "SVG diagram" tools: the outer file is `.svg`, but each icon is `<image href="data:image/png;base64,...">` or `<image href="./icons/lambda.png">`. Opening it in Illustrator/Figma/Inkscape gives you an image object you can move and resize, not a shape you can recolor, reshape, or edit a node of. This skill's icon pipeline (`scripts/stencil2svg.py`, `scripts/svg_inline.py`) exists specifically to avoid this — every icon is real vector geometry (`<path>`, `<rect>`, `<ellipse>`) converted or extracted from the source, never a raster reference.
+**1. Raster icons wrapped in an SVG shell.** The most common failure mode among "SVG diagram" tools: the outer file is `.svg`, but each icon is `<image href="data:image/png;base64,...">` or `<image href="./icons/lambda.png">`. Opening it in Illustrator/Figma/Inkscape gives you an image object you can move and resize, not a shape you can recolor, reshape, or edit a node of. This skill's icon pipeline (`engine/stencil2svg.py`, `engine/svg_inline.py`) exists specifically to avoid this — every icon is real vector geometry (`<path>`, `<rect>`, `<ellipse>`) converted or extracted from the source, never a raster reference.
 
-**2. `<use>` clones instead of inlined geometry.** The obvious way to avoid repeating the same icon's markup for every node that uses it is `<defs><symbol id="lambda">...</symbol></defs>` plus `<use href="#lambda">` per node — smaller file, same icon defined once. This looks editable (Figma imports it, Inkscape shows it in the Symbols dialog) but isn't, fully: both Inkscape's and the SVG spec's own semantics treat a `<use>` clone's contents as **not independently editable** — you can move, scale, or recolor the whole clone, but you cannot select and reshape one path inside it without first unlinking it ("Symbol to Group" in Inkscape) back into real geometry. `render.py` inlines each icon's body directly into every node's `<g>` instead. The cost is a larger file (each copy of an icon repeats its full path data); the benefit is that every single icon instance, on every node, is independently, fully editable the moment the file opens — no unlink step required.
+**2. `<use>` clones instead of inlined geometry.** The obvious way to avoid repeating the same icon's markup for every node that uses it is `<defs><symbol id="lambda">...</symbol></defs>` plus `<use href="#lambda">` per node — smaller file, same icon defined once. This looks editable (Figma imports it, Inkscape shows it in the Symbols dialog) but isn't, fully: both Inkscape's and the SVG spec's own semantics treat a `<use>` clone's contents as **not independently editable** — you can move, scale, or recolor the whole clone, but you cannot select and reshape one path inside it without first unlinking it ("Symbol to Group" in Inkscape) back into real geometry. `engine/render.py` inlines each icon's body directly into every node's `<g>` instead. The cost is a larger file (each copy of an icon repeats its full path data); the benefit is that every single icon instance, on every node, is independently, fully editable the moment the file opens — no unlink step required.
 
 **3. Text converted to outlines.** Exporting text as paths ("convert to outlines") guarantees pixel-identical rendering regardless of font availability, which is why some diagram tools do it by default. It also means the text is no longer text — you cannot select it, search it, or retype it; it's just more vector shapes. `render.py` never does this. Every label is a real `<text>` element with the actual string content. Reopening the SVG anywhere shows real, selectable, retypeable, searchable text — the tradeoff is that a viewer without the referenced font falls back to their system default, which is an acceptable cost for keeping the text real.
