+++ skills/architecture-diagram/references/icons-generic.md
@@ -4,7 +4,7 @@ For `provider: generic` nodes, or any component with no cloud-provider equivalen
 
 All icons: `viewBox="0 0 64 64"`, render at 44×44px default. Stroke `currentColor`, `stroke-width="2.5"` primary shapes. Fill `currentColor` with `fill-opacity` 0.08–0.2 for enclosed shapes.
 
-Use the slug below each heading as the node's `service` value with `provider: generic` (or set `provider: generic` at the top level and omit it per node). `scripts/render.py` resolves it through `assets/generic-icons.json` — no network fetch, no cache directory, always available. A generic-provider node with no `service` set (or an unrecognized slug) still renders cleanly as a labeled colored placeholder; reach for a specific glyph here when it adds real clarity over that placeholder.
+Use the slug below each heading as the node's `service` value with `provider: generic` (or set `provider: generic` at the top level and omit it per node). `engine/render.py` resolves it through `assets/generic-icons.json` — no network fetch, no cache directory, always available. A generic-provider node with no `service` set (or an unrecognized slug) still renders cleanly as a labeled colored placeholder; reach for a specific glyph here when it adds real clarity over that placeholder.
 
 ## Catalog
